Disable mobile view on popups
This sets the container width on the widget, which should cause the iframe
around the widget to be resized to at least 500 pixels. This will mean that
none of the mobile media queries will apply, disabling the mobile view for the
widget.

The motivation for this was the inconsistent display of the popup. Sometimes the
mobile view would be displayed and sometimes the desktop view would be. This is
due to the width dependencies of the iframe and the widget. The iframe is set up
to automatically adjust to the width of the widget, which is based on the width
of the container (in this case the iframe) in the mobile view (less than 500px).
So if the container starts out at a smaller size, this would subsequently cause
the widget to resize itself to 100% of that smaller size, and no further
resizing would occur, resulting in a mobile view being displayed in a
desktop-size window.

One solution to this problem may be to size the iframe to 100% of the browser
window, creating a large enough container for the widget.

This is just a temporary solution until we find a way to allow responsive styles
on the popup.